requirements.txt
asn1crypto==0.23.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
ChatterBot==0.7.6
chatterbot-corpus==1.0.1
click==6.7
cmsplugin-filer==1.1.3
configparser==3.5.0
cryptography==2.1.3
dj-database-url==0.4.2
Django==1.8.18
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==1.3
django-sekizai==0.10.0
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-installer==0.9.7
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.5
enum34==1.1.6
Flask==0.12.2
functools32==3.2.3.post2
future==0.16.0
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
jsondatabase==0.1.7
JsonForm==0.0.2
jsonschema==2.6.0
JsonSir==0.0.2
lxml==4.1.0
MarkupSafe==1.0
nltk==3.2.5
oauthlib==2.0.6
olefile==0.44
pdfkit==0.6.1
Pillow==4.3.0
psycopg2==2.7.3.2
pyasn1==0.3.7
pycparser==2.18
pymongo==3.5.1
pyOpenSSL==17.3.0
PyPDF2==1.26.0
python-dateutil==2.6.1
Python-EasyConfig==0.1.7
python-twitter==3.3
pytz==2017.3
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
Resource==0.2.1
rsa==3.4.2
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.0
singledispatch==3.4.0.3
six==1.11.0
South==1.0.2
SQLAlchemy==1.1.14
tornado==4.5.2
tzlocal==1.4
Unidecode==0.4.21
urllib3==1.22
virtualenv==15.1.0
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.1
xhtml2pdf==0.0.6
